add tests for JpaExceptionHandler
